refactor: move the router provider to index file, and remove home file passing it content to App file
